Extract formatDuration and formatTokens to shared formatters module
