+++ crates/ensemble-ui/src-ui/src/components/RunningTable.tsx
@@ -2,25 +2,10 @@ import { Link } from "react-router-dom";
 import type { RunningSessionRow } from "@/generated/models";
 import { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from "@/components/ui/table";
 import StatusBadge from "./StatusBadge";
+import { formatDuration, formatTokens } from "@/lib/formatters";
 
 interface RunningTableProps {
   sessions: RunningSessionRow[];
-}
-
-function formatDuration(startedAt: string): string {
-  const ms = Date.now() - new Date(startedAt).getTime();
-  const seconds = Math.floor(ms / 1000);
-  if (seconds < 60) return `${seconds}s`;
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m ${seconds % 60}s`;
-  const hours = Math.floor(minutes / 60);
-  return `${hours}h ${minutes % 60}m`;
-}
-
-function formatTokens(n: number): string {
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
-  return String(n);
 }
 
 export default function RunningTable({ sessions }: RunningTableProps) {
