+++ crates/ensemble-ui/src-ui/src/lib/formatters.ts
@@ -1,0 +1,23 @@
+/**
+ * Format a duration in milliseconds to human-readable string
+ * Examples: "45s", "2m 30s", "1h 15m"
+ */
+export function formatDuration(startedAt: string): string {
+  const ms = Date.now() - new Date(startedAt).getTime();
+  const seconds = Math.floor(ms / 1000);
+  if (seconds < 60) return `${seconds}s`;
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes}m ${seconds % 60}s`;
+  const hours = Math.floor(minutes / 60);
+  return `${hours}h ${minutes % 60}m`;
+}
+
+/**
+ * Format token count with k/M suffix
+ * Examples: "500", "1.2k", "1.5M"
+ */
+export function formatTokens(n: number): string {
+  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
+  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
+  return String(n);
+}
